+++ tests/Datasets/LabeledTest.php
@@ -700,4 +700,25 @@ class LabeledTest extends TestCase
             
         $this->assertEquals($expected, $this->dataset->toCsv());
     }
+
+    public function test_to_string() : void
+    {
+        $expected = PHP_EOL
+            . '| Column 0    | Column 1    | Column 2    | Column 3    | Label       |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL
+            . '| nice        | furry       | friendly    | 4           | not monster |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL
+            . '| mean        | furry       | loner       | -1.5        | monster     |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL
+            . '| nice        | rough       | friendly    | 2.6         | not monster |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL
+            . '| mean        | rough       | friendly    | -1          | monster     |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL
+            . '| nice        | rough       | friendly    | 2.9         | not monster |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL
+            . '| nice        | furry       | loner       | -5          | not monster |' . PHP_EOL
+            . '-----------------------------------------------------------------------' . PHP_EOL;
+
+        $this->assertEquals($expected, (string) $this->dataset);
+    }
 }