+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Removed range() from the activation function interface
     - Added label type validation for supervised learners
     - Added toArray, toJson, and toCsv methods to Dataset API
+    - Can now preview a Dataset object in console by echoing it
 
 - 0.0.17-beta
     - Added Tensor extension compatibility